[refactor] Convert trial_response into subfeature
